+++ kurtosis-module/impl/module_io/default_params.go
@@ -1,14 +1,54 @@
 package module_io
 
+const (
+	// If these values are provided for the EL/CL images, then the client type-specific default image will be used
+	useDefaultElImageKeyword = ""
+	useDefaultClImageKeyword = ""
+)
+
+var defaultElImages = map[ParticipantELClientType]string{
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	ParticipantELClientType_Geth: "parithoshj/geth:merge-f72c361", // From around 2022-01-18
+	ParticipantELClientType_Nethermind: "nethermindeth/nethermind:kintsugi_0.5",
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+}
+
+var defaultClImages = map[ParticipantCLClientType]string{
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	ParticipantCLClientType_Lighthouse: "sigp/lighthouse:latest-unstable",
+	ParticipantCLClientType_Teku:       "consensys/teku:latest",
+	ParticipantCLClientType_Nimbus:     "statusim/nimbus-eth2:amd64-latest",
+	// NOTE: Prysm actually has two images - a Beacon and a validator - so we pass in a comma-separated
+	//  "beacon_image,validator_image" string
+	ParticipantCLClientType_Prysm:      "prysmaticlabs/prysm-beacon-chain:latest,prysmaticlabs/prysm-validator:latest",
+	ParticipantCLClientType_Lodestar:   "chainsafe/lodestar:next",
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+}
+
+// To see the exact JSON keys needed to override these values, see the ExecuteParams object and look for the
+//  `json:"XXXXXXX"` metadata on the ExecuteParams properties
 func GetDefaultExecuteParams() *ExecuteParams {
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 	return &ExecuteParams{
 		Participants: []*ParticipantParams{
 			{
-				ELClientType: ParticipantELClientType_Geth,
-				CLClientType: ParticipantCLClientType_Nimbus,
+				ELClientType:  ParticipantELClientType_Geth,
+				ELClientImage: useDefaultElImageKeyword,
+				CLClientType:  ParticipantCLClientType_Nimbus,
+				CLClientImage: useDefaultClImageKeyword,
 			},
 		},
-		Network:             &NetworkParams{
+		Network: &NetworkParams{
 			NetworkID:                          "3151908",
 			DepositContractAddress:             "0x4242424242424242424242424242424242424242",
 			SecondsPerSlot:                     12,
@@ -19,7 +59,11 @@ func GetDefaultExecuteParams() *ExecuteParams {
 			NumValidatorKeysPerNode:            64,
 			PreregisteredValidatorKeysMnemonic: "giant issue aisle success illegal bike spike question tent bar rely arctic volcano long crawl hungry vocal artwork sniff fantasy very lucky have athlete",
 		},
+		WaitForMining:       true,
 		WaitForFinalization: false,
 		ClientLogLevel:      ParticipantLogLevel_Info,
 	}
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 }