+++ kurtosis-module/impl/module_io/params.go
@@ -3,24 +3,38 @@ package module_io
 // Participant log level "enum"
 type ParticipantLogLevel string
 const (
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 	ParticipantLogLevel_Error ParticipantLogLevel = "error"
 	ParticipantLogLevel_Warn ParticipantLogLevel = "warn"
 	ParticipantLogLevel_Info  ParticipantLogLevel = "info"
 	ParticipantLogLevel_Debug ParticipantLogLevel = "debug"
+	ParticipantLogLevel_Trace ParticipantLogLevel = "trace"
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 )
 var validParticipantLogLevels = map[ParticipantLogLevel]bool{
 	ParticipantLogLevel_Error: true,
 	ParticipantLogLevel_Warn:  true,
 	ParticipantLogLevel_Info:  true,
 	ParticipantLogLevel_Debug: true,
+	ParticipantLogLevel_Trace: true,
 }
 
 // Participant EL client type "enum"
 type ParticipantELClientType string
 const (
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 	ParticipantELClientType_Geth       ParticipantELClientType = "geth"
 	ParticipantELClientType_Nethermind ParticipantELClientType = "nethermind"
 	ParticipantELClientType_Besu       ParticipantELClientType = "besu"
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 )
 var validParticipantELClientTypes = map[ParticipantELClientType]bool{
 	ParticipantELClientType_Geth:       true,
@@ -31,11 +45,17 @@ var validParticipantELClientTypes = map[ParticipantELClientType]bool{
 // Participant CL client type "enum"
 type ParticipantCLClientType string
 const (
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 	ParticipantCLClientType_Lighthouse ParticipantCLClientType = "lighthouse"
 	ParticipantCLClientType_Teku       ParticipantCLClientType = "teku"
 	ParticipantCLClientType_Nimbus     ParticipantCLClientType = "nimbus"
 	ParticipantCLClientType_Prysm      ParticipantCLClientType = "prysm"
 	ParticipantCLClientType_Lodestar   ParticipantCLClientType = "lodestar"
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	//       If you change these in any way, modify the example JSON config in the README to reflect this!
+	// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 )
 var validParticipantCLClientTypes = map[ParticipantCLClientType]bool{
 	ParticipantCLClientType_Lighthouse: true,
@@ -52,6 +72,12 @@ type ExecuteParams struct {
 	// Parameters controlling the settings of the network itself
 	Network *NetworkParams	`json:"network"`
 
+	// If set to false, we won't wait for the EL clients to mine at least 1 block before proceeding with adding the CL clients
+	// This is purely for debug purposes; waiting for blockNumber > 0 is required for the CL network to behave as
+	//  expected, but that wait can be several minutes. Skipping the wait can be a good way to shorten the debug loop on a
+	//  CL client that's failing to start.
+	WaitForMining bool			`json:"waitForMining"`
+
 	// If set, the module will block until a finalized epoch has occurred
 	WaitForFinalization bool	`json:"waitForFinalization"`
 
@@ -60,8 +86,19 @@ type ExecuteParams struct {
 }
 
 type ParticipantParams struct {
-	ELClientType ParticipantELClientType `json:"el"`
-	CLClientType ParticipantCLClientType `json:"cl"`
+	// The type of EL client that should be started
+	ELClientType ParticipantELClientType `json:"elType"`
+
+	// The Docker image that should be used for the EL client; leave blank to use the default
+	ELClientImage string				 `json:"elImage"`
+
+	// The type of CL client that should be started
+	CLClientType ParticipantCLClientType `json:"clType"`
+
+	// The Docker image that should be used for the EL client; leave blank to use the default
+	// NOTE: Prysm is different in that it requires two images - a Beacon and a validator
+	//  For Prysm and Prysm only, this field should contain a comma-separated string of "beacon_image,validator_image"
+	CLClientImage string				 `json:"clImage"`
 }
 
 // Parameters controlling particulars of the Eth1 & Eth2 networks
@@ -97,6 +134,5 @@ type NetworkParams struct {
 
 	// This menmonic will a) be used to create keystores for all the types of validators that we have and b) be used to generate a CL genesis.ssz that has the children
 	//  validator keys already preregistered as validators
-	// preregisteredValidatorKeysMnemonic = "giant issue aisle success illegal bike spike question tent bar rely arctic volcano long crawl hungry vocal artwork sniff fantasy very lucky have athlete"
 	PreregisteredValidatorKeysMnemonic string	`json:"preregisteredValidatorKeysMnemonic"`
 }